bot/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
elastic-transport==8.15.0
elasticsearch==8.15.1
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.14.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.3
yarl==1.9.4
